+++ redpipe/fields.py
@@ -13,12 +13,7 @@ import json
 import re
 import typing
 
-# python 3.7 compatibility change
-from typing import (TypeVar, Generic, Optional, Union)
-try:
-    from typing import Protocol
-except ImportError:
-    from typing_extensions import Protocol  # type: ignore
+from typing import (TypeVar, Generic, Optional, Union, Protocol)
 
 
 from .exceptions import InvalidValue